+++ features/support/env.rb
@@ -40,6 +40,10 @@ Before('@skip_below_android_11') do |scenario|
   skip_this_scenario("Skipping scenario") if Maze.config.os_version < 11
 end
 
+Before('@skip_below_android_12') do |scenario|
+  skip_this_scenario("Skipping scenario") if Maze.config.os_version < 12
+end
+
 Before('@skip_below_android_9') do |scenario|
   skip_this_scenario("Skipping scenario") if Maze.config.os_version < 9
 end